# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OrderItem {
  id: string;
  name: string;
  price: string;
  quantity: number;
  image?: string | null;
}

export interface CreateOrderRequest {
  customerName?: string;
  customerPhone?: string;
  items: OrderItem[];
  total: string;
  notes?: string;
}

export type UpdateOrderStatusRequestStatus = typeof UpdateOrderStatusRequestStatus[keyof typeof UpdateOrderStatusRequestStatus];


export const UpdateOrderStatusRequestStatus = {
  pending: 'pending',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName?: string | null;
  customerPhone?: string | null;
  items: OrderItem[];
  total: string;
  notes?: string | null;
  status: OrderStatus;
  createdAt: string;
}