# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName?: string | null;
  customerPhone?: string | null;
  items: OrderItem[];
  total: string;
  notes?: string | null;
  status: OrderStatus;
  createdAt: Date;
}